Add new API endpoints and fix data handling issues in ML analysis
Update datetime handling in advanced_ml_router.py, add new API endpoints for outbreak status, training, embeddings, and governance stats, and fix data type conversions in consent_extraction.py and governance.py.

+++ python_backend/ml_analysis/consent_extraction.py
@@ -228,7 +228,7 @@ class DifferentialPrivacy:
         max_val: float
     ) -> float:
         """Privatize a mean value"""
-        true_mean = np.mean(values)
+        true_mean = float(np.mean(values))
         sensitivity = (max_val - min_val) / len(values)
         return self.add_laplace_noise(true_mean, sensitivity)
